backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
prophet==1.1.5
